Now shows time currently logged not including those currently clocking time.

+++ Models/ProjectLogViewModel.cs
@@ -9,6 +9,6 @@
         public List<TblLaborTimeStamp> LaborTimeStamps { get; set; }
 
         // Property to hold amount of time worked logged on this project minus work currently being clocked
-        public int HoursWorked { get; set; }
+        public double HoursWorked { get; set; }
     }
 }
